GroupHotelSorter

+++ backend/hackaton-engine/hackaton-engine/Models/Preference.cs
@@ -7,7 +7,7 @@ namespace hackaton_engine.Models
         public Tags[] Tags { get; set; }
         public Localizations[] Localizations { get; set; }
         public MustHaves[] MustHaves { get; set; }
-        public Tuple<decimal, decimal> PriceRange { get; set; }
+        public Tuple<double, double> PriceRange { get; set; }
         public Tuple<DateTime, DateTime> DateRange { get; set; }
     }
 }
